requirements.txt
# requirements.txt
fastapi
uvicorn
mysql-connector-python
flask
# EOF